+++ lib/blocks.py
@@ -820,7 +820,7 @@ def list_tx (db, block_hash, block_index, block_time, tx_hash, tx_index):
     cursor.close()
     return
 
-def follow (db, zeroconf):
+def follow (db, mempool):
     # TODO: This is not thread-safe!
     cursor = db.cursor()
 
@@ -851,7 +851,7 @@ def follow (db, zeroconf):
     else:
         tx_index = 0
 
-    zeroconf_transactions_seen = []
+    mempool_transactions_seen = []
     while True:
 
         # Get new blocks.
@@ -923,36 +923,36 @@ def follow (db, zeroconf):
 
         else:
             # Log and ephemerally parse transactions in Bitcoin Core mempool.
-            if zeroconf:
+            if mempool:
                 try:
                     with db:
                         # Fake values for fake block.
                         curr_time = time.time()
-                        zeroconf_tx_index = tx_index
+                        mempool_tx_index = tx_index
 
                         # List the fake block.
                         cursor.execute('''INSERT INTO blocks(
                                             block_index,
                                             block_hash,
                                             block_time) VALUES(?,?,?)''',
-                                            (config.ZEROCONF_BLOCK_INDEX,
-                                             config.ZEROCONF_BLOCK_HASH,
+                                            (config.MEMPOOL_BLOCK_INDEX,
+                                             config.MEMPOOL_BLOCK_HASH,
                                              curr_time)
                                       )
 
                         # List zero‐confirmation transactions.
                         for tx_hash in bitcoin.get_mempool():
-                            list_tx(db, config.ZEROCONF_BLOCK_HASH, config.ZEROCONF_BLOCK_INDEX, curr_time, tx_hash, zeroconf_tx_index)
-                            zeroconf_tx_index += 1
+                            list_tx(db, config.MEMPOOL_BLOCK_HASH, config.MEMPOOL_BLOCK_INDEX, curr_time, tx_hash, MEMPOOL_tx_index)
+                            mempool_tx_index += 1
 
                         # Parse zero‐confirmation transactions.
                         cursor.execute('''SELECT * FROM transactions \
                                           WHERE block_index = ?''',
-                                       (config.ZEROCONF_BLOCK_INDEX,))
+                                       (config.MEMPOOL_BLOCK_INDEX,))
                         for tx in list(cursor):
-                            if tx['tx_hash'] not in zeroconf_transactions_seen:
+                            if tx['tx_hash'] not in mempool_transactions_seen: # If not already logged as zero‐conf…
                                 parse_tx(db, tx)
-                                zeroconf_transactions_seen.append(tx['tx_hash'])
+                                mempool_transactions_seen.append(tx['tx_hash'])
 
                         assert False    # Never commit!
                 except AssertionError: