+++ lib/config.py
@@ -6,7 +6,7 @@ UNIT = 100000000        # The same across currencies.
 # Versions
 VERSION_MAJOR = 9
 VERSION_MINOR = 26
-VERSION_REVISION = 0
+VERSION_REVISION = 1
 VERSION_STRING = str(VERSION_MAJOR) + '.' + str(VERSION_MINOR) + '.' + str(VERSION_REVISION)
 
 # Counterparty protocol
@@ -22,9 +22,11 @@ MEMPOOL_STATUS = 'MEMPOOL'
 # SQLite3
 MAX_INT = 2**63 - 1
 
+# Bitcoin Core
+OP_RETURN_MAX_SIZE = 40 # bytes
+
 # Protocol defaults
 # NOTE: If the DUST_SIZE constants are changed, they MUST also be changed in counterblockd/lib/config.py as well
-ENCODING = 'multisig'
 REGULAR_DUST_SIZE = 5430        # TODO: This is just a guess. I got it down to 5530 satoshis.
 MULTISIG_DUST_SIZE = 5430 * 2   # TODO: This is just a guess. I did it down to 1.4x. (Used for regular outputs in multi‐sig transactions, too.)
 OP_RETURN_VALUE = 0